+++ src/config/ci.php
@@ -118,7 +118,7 @@ return [
         'notifier' => 'PragmaRX\TestsWatcher\Notifications',
     ],
 
-    /*
+    /**
      * Editors
      *
      */
@@ -126,7 +126,7 @@ return [
         'phpstorm' => [
             'type' => 'PHPStorm',
 
-            'bin' => '/usr/local/bin/pstorm',
+            'bin' => '/usr/local/bin/pstorm {$file} --line {$line}',
 
             'default' => true,
         ],
@@ -134,13 +134,13 @@ return [
         'sublime' => [
             'type' => 'SublimeText 3',
 
-            'bin' => '/usr/local/bin/subl',
+            'bin' => '/usr/local/bin/subl {$file}:{$line}',
         ],
 
         'vscode' => [
             'type' => 'VSCode',
 
-            'bin' => '/Applications/Visual\ Studio\ Code.app/Contents/Resources/app/bin/code',
+            'bin' => '/Applications/Visual\ Studio\ Code.app/Contents/Resources/app/bin/code --goto {$file}:{$line}',
         ],
     ],