+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Config for PHPStorm editor
 - Config for Sublime Text 3 editor
 - Config for Visual Studio Code editor
+### Changed
+- Allow better configuration of editor's binary
 ### Fixed
 - Abending when tester used in suite does not exists
 - Piper script not being